Assistant checkpoint: Refactored job duplication logic into a dedicated service
Assistant generated file changes:
- lib/services/job-duplication.service.ts: Create new job duplication service
- lib/services/pi-job-mapping.service.ts: Add method to get mappings by job ID, Add method to get mappings by job ID
- app/api/jobs/duplicate/route.ts: Refactor route to use the new JobDuplicationService

---

User prompt:

Refactor the code in the @app/api/jobs/duplicate/route.ts so that the actual job duplication logic is a new service file in `lib/services`
